+++ src/main/java/com/ilcle/ilcle_back/repository/PostRepositoryCustom.java
@@ -1,6 +1,8 @@
 package com.ilcle.ilcle_back.repository;
 
 import com.ilcle.ilcle_back.dto.response.PostResponseDto;
+import com.ilcle.ilcle_back.entity.Member;
+import com.ilcle.ilcle_back.entity.Post;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
@@ -8,4 +10,8 @@ public interface PostRepositoryCustom {
 
 	// 전체글 조회(최신순)
 	Page<PostResponseDto> getAllPosts(Pageable pageable);
+
+	Page<Post> findFilterByMember(Member member, Pageable pageable, Boolean read);
+
+
 }
